+++ internal/api/dedup.go
@@ -1,0 +1,52 @@
+package api
+
+import (
+	"encoding/json"
+	"log"
+
+	"github.com/aceberg/unbox/internal/check"
+)
+
+func deduplicate() {
+
+	config := getConfigFromFile()
+
+	outbounds, ok := config["outbounds"].([]any)
+	if !ok {
+		log.Println("No outbounds found in", Config.OutPath)
+		return
+	}
+
+	seen := make(map[string]struct{})
+	var unique []ProxyServer
+
+	for _, ob := range outbounds {
+		m, ok := ob.(map[string]any)
+		if !ok {
+			continue
+		}
+
+		// Copy without tag
+		cmp := make(map[string]any, len(m))
+		for k, v := range m {
+			if k != "tag" {
+				cmp[k] = v
+			}
+		}
+
+		keyBytes, err := json.Marshal(cmp)
+		check.IfError(err)
+
+		key := string(keyBytes)
+
+		if _, exists := seen[key]; exists {
+			log.Println("Dup", m["tag"].(string))
+			continue // duplicate except for tag
+		}
+
+		seen[key] = struct{}{}
+		unique = append(unique, ProxyServer{Tag: m["tag"].(string)})
+	}
+
+	editConfig(unique)
+}